Read CPU definition from config file.
The CPU definition is now completely contained in one configuration
file and read on instantiation of the `Program` class. This ensures
that a program cannot leak information between multiple instantiations.
See also `tests/test_isolation.py`. Note that currently a program does
not modify its state, but this can be reasonable (for example to track
the contents of certain registers to enable certain peephole
optimizations).